+++ src/lib.rs
@@ -83,6 +83,7 @@ pub use xum1541::DeviceAccessError;
 pub use xum1541::DeviceChannel;
 pub use xum1541::Error as Xum1541Error;
 pub use xum1541::{Device, RemoteUsbDevice, UsbDevice};
+pub use xum1541::DeviceInfo as Xum1541DeviceInfo;
 
 pub use xum1541::device::remoteusb::DEFAULT_ADDR as DEFAULT_REMOTE_ADDR;
 pub use xum1541::device::remoteusb::DEFAULT_PORT as DEFAULT_REMOTE_PORT;