+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.3.2] - 2025-??-??
+### Changed
+- Moved examples/cli to bin/cli
+- Added support to bin/cli for serial numbers
+- Added cbm.xum1541_info() to get xum1541 device info
+
 ## [0.3.1] - 2025-02-08
 ### Changed
 - Move to xum1541 0.3.1